+++ core/arch/freeRTOS/include/forte/CMakeLists.txt
@@ -1,0 +1,15 @@
+#############################################################################
+# Copyright (c) 2025 Martin Erich Jobst
+#
+# This program and the accompanying materials are made available under the
+# terms of the Eclipse Public License 2.0 which is available at
+# http://www.eclipse.org/legal/epl-2.0.
+#
+# SPDX-License-Identifier: EPL-2.0
+#
+# Contributors:
+#    Martin Erich Jobst
+#      - initial API and implementation and/or initial documentation
+#############################################################################
+
+add_subdirectory(arch)
