+++ core/arch/freeRTOS/src/forte_architecture_time.cpp
@@ -12,8 +12,8 @@
  *  Martin Jobst - add high-resolution realtime clock fallback
  *******************************************************************************/
 
-#include <FreeRTOS.h>
-#include <task.h>
+#include <freertos/FreeRTOS.h>
+#include <freertos/task.h>
 
 #include "forte/arch/forte_architecture_time.h"
 #include "forte/util/forte_constants.h"
@@ -24,18 +24,18 @@ uint_fast64_t getNanoSecondsMonotonicArch() {
 }
 
 uint_fast64_t getNanoSecondsRealtimeArch() {
-  return static_cast<uint_fast64_t>(forte_time()) * 1000000000LL;
+  return static_cast<uint_fast64_t>(forte::arch::forte_time()) * 1000000000LL;
 }
 
-time_t forte_timegm(struct tm *pa_tm) {
+time_t forte::arch::forte_timegm(struct tm *pa_tm) {
   time_t tm = mktime(pa_tm);
   return mktime(gmtime(&tm));
 }
 
-struct tm *forte_localtime(const time_t *paTime, struct tm *const paResult) {
+struct tm *forte::arch::forte_localtime(const time_t *paTime, struct tm *const paResult) {
   return localtime_r(paTime, paResult);
 }
 
-struct tm *forte_gmtime(const time_t *const paTime, struct tm *const paResult) {
+struct tm *forte::arch::forte_gmtime(const time_t *const paTime, struct tm *const paResult) {
   return gmtime_r(paTime, paResult);
 }
